refactor: updated logic to be more modular
