+++ src/backend/python/config.py
@@ -1,14 +1,15 @@
 import os
 from dotenv import load_dotenv
 
-# Cargar variables de entorno
+# Load environment variables
 load_dotenv()
 
-# Configuración de los modelos
+# Model configuration
 MODEL_GPT = "gpt-4o-mini"
 MODEL_LLAMA = "llama3.2"
 
-# Claves API
+# API Keys
 OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
+
 if not OPENAI_API_KEY:
-    raise ValueError("Falta la clave OPENAI_API_KEY en el archivo .env")
+    raise ValueError("❌ Missing OPENAI_API_KEY in the .env file")
